+++ src/main/java/warmup/one/WarmupOneFourthSet.java
@@ -18,4 +18,22 @@ public class WarmupOneFourthSet {
 
         return str.substring(0, n) + str.substring(n+1);
     }
+
+    /**
+     * Given a string, return a new string where the first and last chars have been exchanged.
+     *
+     * @param str
+     * @return
+     */
+    public String frontBack(String str) {
+        if (str.length() == 0 || str.length() == 1) {
+            return str;
+        }
+
+        String firstCharacter = str.substring(0, 1);
+        String lastCharacter = str.substring(str.length()-1);
+        String otherCharacters = str.substring(1, str.length()-1);
+
+        return lastCharacter + otherCharacters + firstCharacter;
+    }
 }