+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.0.1] - 2022-09-25
 ### Added
 - Warmup-1 missingChar
+- Warmup-1 frontBack
 
 ## [1.0.0] - 2022-09-24
 ### Added